[wasm] Make the wasm sdk not depends on desktop sdk.
The issue was that packager.exe was using Mono.Cecil from net_4_x, which is only built/installed if the desktop sdk is built. We replace it with a csproj that fetches Cecil from nuget instead.

Finally, we change the CI script to disable desktop as well to ensure we don't regress in that scenario.